Use Project and Task serializers everywhere
Previously, we were only running Projects and Tasks through our serializers in
the Index actions. There, we are explicitly calling the serializer classes.
In the other controller actions, Rails was looking for serializers in the
app/serializers directory, not app/serializers/v1. Relocating them
renders consistent JSON in our different endpoints
