+++ app/serializers/project_serializer.rb
@@ -1,0 +1,7 @@
+class ProjectSerializer < ActiveModel::Serializer
+
+  attributes :id, :name, :description, :state
+
+  has_many :tasks
+
+end
