+++ app/serializers/task_serializer.rb
@@ -1,0 +1,7 @@
+class TaskSerializer < ActiveModel::Serializer
+
+  attributes :id, :name, :description, :state
+
+  belongs_to :project_id
+
+end
