--- app/serializers/v1/project_serializer.rb
@@ -1,8 +1,0 @@
-module V1
-  class ProjectSerializer < ActiveModel::Serializer
-
-    attributes :id, :name, :description, :state
-
-    has_many :tasks
-  end
-end
