--- app/serializers/v1/task_serializer.rb
@@ -1,8 +1,0 @@
-module V1
-  class TaskSerializer < ActiveModel::Serializer
-
-    attributes :id, :name, :description, :state
-
-    belongs_to :project_id
-  end
-end
